+++ modules/gltf/editor/editor_scene_exporter_gltf_plugin.h
@@ -31,7 +31,8 @@
 #ifndef EDITOR_SCENE_EXPORTER_GLTF_PLUGIN_H
 #define EDITOR_SCENE_EXPORTER_GLTF_PLUGIN_H
 
-#if TOOLS_ENABLED
+#ifdef TOOLS_ENABLED
+
 #include "editor/editor_plugin.h"
 #include "editor_scene_importer_gltf.h"
 
@@ -47,5 +48,7 @@ public:
 	bool has_main_screen() const override;
 	SceneExporterGLTFPlugin();
 };
+
 #endif // TOOLS_ENABLED
+
 #endif // EDITOR_SCENE_EXPORTER_GLTF_PLUGIN_H
