+++ modules/gridmap/SCsub
@@ -5,4 +5,7 @@ Import("env_modules")
 
 env_gridmap = env_modules.Clone()
 
+# Godot's own source files
 env_gridmap.add_source_files(env.modules_sources, "*.cpp")
+if env["tools"]:
+    env_gridmap.add_source_files(env.modules_sources, "editor/*.cpp")
