+++ modules/gridmap/register_types.cpp
@@ -28,21 +28,25 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#include "register_types.h"
 #ifndef _3D_DISABLED
+
+#include "register_types.h"
+
 #include "core/object/class_db.h"
 #include "grid_map.h"
-#include "grid_map_editor_plugin.h"
+
+#ifdef TOOLS_ENABLED
+#include "editor/grid_map_editor_plugin.h"
 #endif
 
 void register_gridmap_types() {
-#ifndef _3D_DISABLED
 	GDREGISTER_CLASS(GridMap);
 #ifdef TOOLS_ENABLED
 	EditorPlugins::add_by_type<GridMapEditorPlugin>();
-#endif
 #endif
 }
 
 void unregister_gridmap_types() {
 }
+
+#endif // _3D_DISABLED
