+++ modules/websocket/SCsub
@@ -41,6 +41,8 @@ elif env["builtin_wslay"]:
 module_obj = []
 
 env_ws.add_source_files(module_obj, "*.cpp")
+if env["tools"]:
+    env_ws.add_source_files(module_obj, "editor/*.cpp")
 env.modules_sources += module_obj
 
 # Needed to force rebuilding the module files when the thirdparty library is updated.
